Upload EDA for Wind - WIP
Uploaded so people can share there opinion and comments on how i am approaching the problem. 

Next Steps -> I will make a table with the Natural Disaster dates and will link those dates to the datapoints and see if there is any pattern, If it works, will train and test the model to predict the Natural Disasters.